Move add_canonical_source_record_id to shared_columns.py
Co-locate with add_canonical_author_columns after PR #130 extraction.
runner.py imports both helpers from shared_columns; behavior unchanged.
