+++ data_platform/preprocessing/shared_columns.py
@@ -6,7 +6,10 @@ from typing import TYPE_CHECKING
 
 import pandas as pd
 
-from data_platform.utils.platform_specific_columns import CANONICAL_AUTHOR_HANDLE_COLUMN
+from data_platform.utils.platform_specific_columns import (
+    CANONICAL_AUTHOR_HANDLE_COLUMN,
+    CANONICAL_SOURCE_RECORD_ID_COLUMN,
+)
 
 if TYPE_CHECKING:
     from data_platform.preprocessing.runner import PreprocessPlatformSpec
@@ -47,4 +50,37 @@ def add_canonical_author_columns(
     source_column = spec.author_handle_source_column
     original_handle = out[source_column]
     out[CANONICAL_AUTHOR_HANDLE_COLUMN] = original_handle.map(lambda value: str(value))
+    return out
+
+
+def add_canonical_source_record_id(
+    df: pd.DataFrame,
+    spec: PreprocessPlatformSpec,
+) -> pd.DataFrame:
+    """Copy the platform's original record id onto the shared ``source_record_id`` column.
+
+    You still have original fields such as Bluesky ``uri``, Reddit
+    ``comment_fullname``, and Twitter ``tweet_id`` on the returned frame. The
+    function does not modify the input frame.
+
+    Parameters
+    ----------
+    spec
+        ``spec.columns.records_id_column`` is the copy source. The destination
+        is always shared ``source_record_id``.
+
+    Returns
+    -------
+    pd.DataFrame
+        A new frame that includes ``source_record_id``.
+
+    Raises
+    ------
+    KeyError
+        When the original record id column is missing from the frame.
+    """
+    out = df.copy()
+    source_column = spec.columns.records_id_column
+    original_id = out[source_column]
+    out[CANONICAL_SOURCE_RECORD_ID_COLUMN] = original_id.map(lambda value: str(value))
     return out
